refractor: move bootstrap to its own workflow
